Implement retry mechanism based on the leader and follower
