+++ tests/integrations/client/http_client_test.go
@@ -438,13 +438,13 @@ func (suite *httpClientTestSuite) TestTransferLeader() {
 	re.NoError(err)
 	re.Len(members.Members, 2)
 
-	oldLeader, err := suite.client.GetLeader(suite.ctx)
+	leader, err := suite.client.GetLeader(suite.ctx)
 	re.NoError(err)
 
 	// Transfer leader to another pd
 	for _, member := range members.Members {
-		if member.Name != oldLeader.Name {
-			err = suite.client.TransferLeader(suite.ctx, member.Name)
+		if member.GetName() != leader.GetName() {
+			err = suite.client.TransferLeader(suite.ctx, member.GetName())
 			re.NoError(err)
 			break
 		}
@@ -453,5 +453,14 @@ func (suite *httpClientTestSuite) TestTransferLeader() {
 	newLeader := suite.cluster.WaitLeader()
 	re.NotEmpty(newLeader)
 	re.NoError(err)
-	re.NotEqual(oldLeader.Name, newLeader)
+	re.NotEqual(leader.GetName(), newLeader)
+	// Force to update the members info.
+	suite.client.(interface{ UpdateMembersInfo() }).UpdateMembersInfo()
+	leader, err = suite.client.GetLeader(suite.ctx)
+	re.NoError(err)
+	re.Equal(newLeader, leader.GetName())
+	members, err = suite.client.GetMembers(suite.ctx)
+	re.NoError(err)
+	re.Len(members.Members, 2)
+	re.Equal(leader.GetName(), members.Leader.GetName())
 }
